LPS-53617 PreviewArticleContentAction: move logic to ActionUtil class (no logic changes)
